% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIRFunctions.R
\name{infect_corners}
\alias{infect_corners}
\title{Initial Matrix With Infected Corners}
\usage{
infect_corners(nrow = 6, ncol = 10)
}
\arguments{
\item{nrow}{Number of Rows}

\item{ncol}{Number of Columns}
}
\value{
SIRmatrix
}
\description{
Initial Matrix With Infected Corners
}
\examples{
infect_corners(5,7)
}
