% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIRFunctions.R
\name{plot.SIRmatrix}
\alias{plot.SIRmatrix}
\title{Plot an SIRMatrix Model}
\usage{
\method{plot}{SIRmatrix}(x, col = c("white", "red", "gray"), ...)
}
\arguments{
\item{x}{Desired SIRmatrix to be plotted}

\item{col}{Desired colors for the plot}

\item{...}{Any other arguments to be passed through the function}
}
\value{
image
}
\description{
Plot an SIRMatrix Model
}
\examples{
B <- infect_corners()
plot(B)
}
