% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIRFunctions.R
\name{random_infect}
\alias{random_infect}
\title{Randomly Infects Cells in a matrix}
\usage{
random_infect(nrow = 6, ncol = 10, prob = 0.1)
}
\arguments{
\item{nrow}{Number of Rows}

\item{ncol}{Number of Columns}

\item{prob}{Probability of Infection}
}
\value{
SIRmatrix
}
\description{
Randomly Infects Cells in a matrix
}
\examples{
random_infect(nrow = 5,ncol = 6,prob = .25)
}
