% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIRFunctions.R
\name{simulate_sir}
\alias{simulate_sir}
\title{Simulate an SIR Model}
\usage{
simulate_sir(x, prob = 0.125)
}
\arguments{
\item{x}{Starting Matrix}

\item{prob}{Probability of Infection}
}
\value{
list
}
\description{
Simulate an SIR Model
}
\examples{
A <- infect_corners()
simulate_sir(A,prob = .125)
}
