% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIRFunctions.R
\name{step}
\alias{step}
\title{Does a Step for the SIR Model}
\usage{
step(x, prob = 0.125)
}
\arguments{
\item{x}{Starting Matrix}

\item{prob}{Probability of Infection}
}
\value{
SIRmatrix
}
\description{
Does a Step for the SIR Model
}
\examples{
A <- infect_corners()
step(A, prob = .125)
}
